feat: Add RNPickerSelect component to Session for selecting ball type
The code changes in the Session component add the RNPickerSelect component to enable the selection of different ball types. This feature allows users to choose from options like "Random," "Bouncer Ball," "Yorker Ball," "Fast Ball," and "Slow Ball." The addition of this component enhances the functionality and user experience of the Session component.